Call CreateIfNotExists(PublicAccessType.None) instead of Blob
The storage account is configured to not have any public access, so we have to use the right API to avoid errors saying that we can't create public blobs in a private account. It isn't clear how/why this ever worked, but this change does reflect the current app/storage settings.